+++ ketch/src/main/java/com/sherafatpour/bluetile/internal/database/DownloadDatabase.kt
@@ -3,7 +3,7 @@ package com.sherafatpour.bluetile.internal.database
 import androidx.room.Database
 import androidx.room.RoomDatabase
 
-@Database(entities = [DownloadEntity::class], version = 5, exportSchema = false)
+@Database(entities = [DownloadEntity::class], version = 6, exportSchema = false)
 internal abstract class DownloadDatabase : RoomDatabase() {
     abstract fun downloadDao(): DownloadDao
 }